net: 6lowpan: Fix 6lowpan fragmentation
IP packet passing through 6lowpan compression and 802.15.4
fragmentation has an issue. (RFC 6282 and 4944) RFC4944, 5.3
says "fragment header's datagram_size and datagram_offset
values as the size and offset of the IPv6 datagram before
compression".

At the moment datagram size and offset values are after
compression. Due to multiple bearer (15.4 and bluetooth) support
and multiple incoming buffers support functionality is divided
to compression and fragmentation parts. And datagram size and
offset values for BT are after compression. Small packet (which
doesn't need fragmentation) doesn't go through this. So cacheing
compressed and uncompressed header lengths and with hdr difference
preparing fragments as per original buffer (total buf size and offset).

Jira: ZEP-208

Former-commit-id: 5e99b14ad4a3871fe6e84b490f3210bd6f628e71

+++ include/net/l2_buf.h
@@ -8,7 +8,7 @@ DECL|l2_buf_unref|macro|l2_buf_unref
 DECL|l2_buf|struct|struct l2_buf {
 DECL|last_tx_status|member|int last_tx_status;
 DECL|packetbuf_hdr_len|member|uint8_t packetbuf_hdr_len;
-DECL|packetbuf_payload_len|member|int packetbuf_payload_len;
+DECL|packetbuf_payload_len|member|uint8_t packetbuf_payload_len;
 DECL|packetbuf_ptr|member|uint8_t *packetbuf_ptr;
 DECL|pkt_buflen|member|uint16_t pkt_buflen, pkt_bufptr;
 DECL|pkt_bufptr|member|uint16_t pkt_buflen, pkt_bufptr;
